+++ src/RegionTrees.jl
@@ -2,7 +2,6 @@ module RegionTrees
 
 using StaticArrays
 import StaticArrays: similar_type, Size
-using IterTools: chain
 import Base: show, size, getindex, convert, parent, @pure
 export Cell,
        split!,