+++ CHANGELOG.md
@@ -15,6 +15,10 @@ Types of changes:
 
 ## [Unreleased]
 
+### Removed
+
+- unused `IterTools.jl` dependency ([#52])
+
 ## [0.3.3] - 2026-07-04
 
 ### Changed
@@ -29,6 +33,7 @@ Types of changes:
 
 [unreleased]: https://github.com/JuliaGeometry/RegionTrees.jl/compare/v0.3.3..HEAD
 [0.3.3]: https://github.com/JuliaGeometry/RegionTrees.jl/releases/tag/v0.3.3
+[#52]: https://github.com/JuliaGeometry/RegionTrees.jl/pull/52
 [#43]: https://github.com/JuliaGeometry/RegionTrees.jl/issues/43
 [#45]: https://github.com/JuliaGeometry/RegionTrees.jl/pull/45
 [#48]: https://github.com/JuliaGeometry/RegionTrees.jl/pull/48